@@ -41,73 +52,73 @@ add_subdirectory(okvis)
 
 
 add_definitions(-fPIC -march=native -O3 -Wextra -Winit-self -Woverloaded-virtual
-                -Wnon-virtual-dtor -Wsign-promo -Wno-long-long -std=c++17)
+        -Wnon-virtual-dtor -Wsign-promo -Wno-long-long -std=c++20)
 
-list(APPEND ament_libraries 
-      rclcpp
-      rosidl_default_generators 
-      pcl_ros 
-      cv_bridge 
-      sensor_msgs 
-      geometry_msgs 
-      tf2
-      tf2_ros 
-      tf2_sensor_msgs 
-      tf2_geometry_msgs
-      std_msgs 
-      nav_msgs 
-      std_srvs
-      visualization_msgs
-      image_transport
-      pcl_conversions
-      )
+list(APPEND ament_libraries
+        rclcpp
+        rosidl_default_generators
+        pcl_ros
+        cv_bridge
+        sensor_msgs
+        geometry_msgs
+        tf2
+        tf2_ros
+        tf2_sensor_msgs
+        tf2_geometry_msgs
+        std_msgs
+        nav_msgs
+        std_srvs
+        visualization_msgs
+        image_transport
+        pcl_conversions
+)
 
-ament_export_dependencies(rosidl_default_runtime )
+ament_export_dependencies(rosidl_default_runtime)
 
 
 add_library(${PROJECT_NAME}_lib STATIC src/Subscriber.cpp
-                               src/Publisher.cpp
-                               src/RosParametersReader.cpp
-                               include/okvis/Subscriber.hpp
-                               include/okvis/Publisher.hpp
-                               include/okvis/RosParametersReader.hpp)
+        src/Publisher.cpp
+        src/RosParametersReader.cpp
+        include/okvis/Subscriber.hpp
+        include/okvis/Publisher.hpp
+        include/okvis/RosParametersReader.hpp)
 add_dependencies(${PROJECT_NAME}_lib okvis_multisensor_processing)
-ament_target_dependencies(${PROJECT_NAME}_lib  ${ament_libraries})
-rosidl_target_interfaces(${PROJECT_NAME}_lib ${PROJECT_NAME} rosidl_typesupport_cpp) 
+ament_target_dependencies(${PROJECT_NAME}_lib ${ament_libraries})
+rosidl_target_interfaces(${PROJECT_NAME}_lib ${PROJECT_NAME} rosidl_typesupport_cpp)
 
 
-target_include_directories( ${PROJECT_NAME}_lib PUBLIC
-  include
-  okvis/okvis_util/include
-  okvis/okvis_kinematics/include
-  okvis/okvis_time/include
-  okvis/okvis_cv/include
-  okvis/okvis_common/include
-  okvis/okvis_ceres/include
-  okvis/okvis_timing/include
-  okvis/okvis_matcher/include
-  okvis/okvis_frontend/include 
-  okvis/okvis_multisensor_processing/include
+target_include_directories(${PROJECT_NAME}_lib PUBLIC
+        include
+        okvis/okvis_util/include
+        okvis/okvis_kinematics/include
+        okvis/okvis_time/include
+        okvis/okvis_cv/include
+        okvis/okvis_common/include
+        okvis/okvis_ceres/include
+        okvis/okvis_timing/include
+        okvis/okvis_matcher/include
+        okvis/okvis_frontend/include
+        okvis/okvis_multisensor_processing/include
 )
 
-target_link_libraries(${PROJECT_NAME}_lib 
-  okvis_util
-  okvis_kinematics
-  okvis_time
-  okvis_cv 
-  okvis_common
-  okvis_ceres
-  okvis_timing
-  okvis_matcher
-  okvis_frontend 
-  okvis_multisensor_processing
-  pthread
+target_link_libraries(${PROJECT_NAME}_lib
+        okvis_util
+        okvis_kinematics
+        okvis_time
+        okvis_cv
+        okvis_common
+        okvis_ceres
+        okvis_timing
+        okvis_matcher
+        okvis_frontend
+        okvis_multisensor_processing
+        pthread
 )
 
 install(TARGETS ${PROJECT_NAME}_lib
-  LIBRARY DESTINATION lib
-  RUNTIME DESTINATION bin
-  PUBLIC_HEADER DESTINATION include
+        LIBRARY DESTINATION lib
+        RUNTIME DESTINATION bin
+        PUBLIC_HEADER DESTINATION include
 )
 install(DIRECTORY include/
         DESTINATION include
