@@ -6,70 +6,63 @@ set(OKVIS_MAJOR_VERSION 1)
 set(OKVIS_MINOR_VERSION 1)
 set(OKVIS_PATCH_VERSION 3)
 set(OKVIS_VERSION
-  ${OKVIS_MAJOR_VERSION}.${OKVIS_MINOR_VERSION}.${OKVIS_PATCH_VERSION})
+        ${OKVIS_MAJOR_VERSION}.${OKVIS_MINOR_VERSION}.${OKVIS_PATCH_VERSION})
 
 # options for build configuration
-option (USE_SYSTEM_BRISK
-        "Use brisk via find_package rather than downloading it as part of okvis" OFF) 
-option (USE_SYSTEM_CERES
-        "Use ceres via find_package rather than downloading it as part of okvis" ON) 
-option (BUILD_APPS
-        "Builds a demo app (which requires boost)" ON) 
-option (BUILD_TESTS
-        "Builds all gtests" OFF) 
+option(USE_SYSTEM_BRISK
+        "Use brisk via find_package rather than downloading it as part of okvis" OFF)
+option(USE_SYSTEM_CERES
+        "Use ceres via find_package rather than downloading it as part of okvis" ON)
+option(BUILD_APPS
+        "Builds a demo app (which requires boost)" ON)
+option(BUILD_TESTS
+        "Builds all gtests" OFF)
 SET(N_CORES 3 CACHE STRING "Using N number of cores for parallel build")
 
 # Offer the user the choice of overriding the installation directories
 set(INSTALL_LIB_DIR lib CACHE PATH "Installation directory for libraries")
 set(INSTALL_BIN_DIR bin CACHE PATH "Installation directory for executables")
 set(INSTALL_INCLUDE_DIR include CACHE PATH
-  "Installation directory for header files")
-if(WIN32 AND NOT CYGWIN)
-  set(DEF_INSTALL_CMAKE_DIR CMake)
-else()
-  set(DEF_INSTALL_CMAKE_DIR lib/CMake/okvis)
-endif()
+        "Installation directory for header files")
+if (WIN32 AND NOT CYGWIN)
+    set(DEF_INSTALL_CMAKE_DIR CMake)
+else ()
+    set(DEF_INSTALL_CMAKE_DIR lib/CMake/okvis)
+endif ()
 set(INSTALL_CMAKE_DIR ${DEF_INSTALL_CMAKE_DIR} CACHE PATH
-  "Installation directory for CMake files")
- 
+        "Installation directory for CMake files")
+
 # Make relative paths absolute (needed later on)
-foreach(p LIB BIN INCLUDE CMAKE)
-  set(var INSTALL_${p}_DIR)
-  if(NOT IS_ABSOLUTE "${${var}}")
-    set(${var} "${CMAKE_INSTALL_PREFIX}/${${var}}")
-  endif()
-endforeach()
+foreach (p LIB BIN INCLUDE CMAKE)
+    set(var INSTALL_${p}_DIR)
+    if (NOT IS_ABSOLUTE "${${var}}")
+        set(${var} "${CMAKE_INSTALL_PREFIX}/${${var}}")
+    endif ()
+endforeach ()
 
 # make sure we use Release and warn otherwise
-if(NOT CMAKE_BUILD_TYPE AND NOT CMAKE_CONFIGURATION_TYPES)
-  message(STATUS "Setting build type to 'Release' as none was specified.")
-  set(CMAKE_BUILD_TYPE Release CACHE STRING "Choose the type of build." FORCE)
-  # Set the possible values of build type for cmake-gui
-  set_property(CACHE CMAKE_BUILD_TYPE PROPERTY STRINGS "Debug" "Release"
-    "MinSizeRel" "RelWithDebInfo")
-endif()
+if (NOT CMAKE_BUILD_TYPE AND NOT CMAKE_CONFIGURATION_TYPES)
+    message(STATUS "Setting build type to 'Release' as none was specified.")
+    set(CMAKE_BUILD_TYPE Release CACHE STRING "Choose the type of build." FORCE)
+    # Set the possible values of build type for cmake-gui
+    set_property(CACHE CMAKE_BUILD_TYPE PROPERTY STRINGS "Debug" "Release"
+            "MinSizeRel" "RelWithDebInfo")
+endif ()
 if (NOT "${CMAKE_BUILD_TYPE}" STREQUAL "Release")
-  message(WARNING "CMAKE_BUILD_TYPE not set to 'Release'. Performance may be terrible.")
-else()
-  message(STATUS "Building with build type '${CMAKE_BUILD_TYPE}'")
-endif()
+    message(WARNING "CMAKE_BUILD_TYPE not set to 'Release'. Performance may be terrible.")
+else ()
+    message(STATUS "Building with build type '${CMAKE_BUILD_TYPE}'")
+endif ()
 
-if(APPLE)
-  # The clang compiler (on osx) is somehow much more strict
-  # than the compilers on ubuntu and so this does not seem
-  # possible on OSX just yet.
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -march=native -Wall -std=c++17 -fPIC")
-else()
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -march=native -Wall -std=c++17 -fPIC")
-endif()
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -march=native -Wall -std=c++20 -fPIC")
 
-if(${CMAKE_HOST_SYSTEM_PROCESSOR} MATCHES "arm*")
-  message(STATUS "ARM processor detected, will attempt to use NEON.")
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -mfpu=neon")
-else()
-  message(STATUS "Assuming SSE instructions available.")
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -mssse3")
-endif()
+if (${CMAKE_HOST_SYSTEM_PROCESSOR} MATCHES "arm*")
+    message(STATUS "ARM processor detected, will attempt to use NEON.")
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -mfpu=neon")
+else ()
+    message(STATUS "Assuming SSE instructions available.")
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -mssse3")
+endif ()
 
 get_directory_property(defs COMPILE_DEFINITIONS)
 
