@@ -6,7 +6,6 @@
 #include <sensor_msgs/msg/compressed_image.hpp>
 #include <sensor_msgs/msg/image.hpp>
 #include <sensor_msgs/msg/imu.hpp>
-
 typedef message_filters::sync_policies::ApproximateTime<sensor_msgs::msg::Image, sensor_msgs::msg::Image>
     image_sync_policy;
 typedef message_filters::sync_policies::ApproximateTime<sensor_msgs::msg::CompressedImage,
@@ -15,13 +14,23 @@ typedef message_filters::sync_policies::ApproximateTime<sensor_msgs::msg::Compre
 
 class StereoSync : public rclcpp::Node {
  public:
-  StereoSync(const rclcpp::NodeOptions& options) : Node("stereo_sync", options) {
+  explicit StereoSync(const rclcpp::NodeOptions& options) : Node("stereo_sync", options) {
     previous_stamp_ = rclcpp::Time(0, 0);
 
     this->get_parameter("left_img_topic", left_img_topic_);
     this->get_parameter("right_img_topic", right_img_topic_);
     this->get_parameter("compressed", compressed_image_);
+    this->get_parameter("config_filename", config_file_);
 
+    cv::FileStorage fsSettings(config_file_, cv::FileStorage::READ);
+
+    if (!fsSettings.isOpened()) {
+      RCLCPP_FATAL_STREAM(this->get_logger(), "ERROR: Wrong path to settings\n");
+    }
+
+    float frame_rate = static_cast<float>(fsSettings["camera_params"]["camera_rate"]);
+    float max_time_diff_between_msgs = 0.5f / frame_rate;
+    RCLCPP_INFO_STREAM(this->get_logger(), "Setting Age Penalty to:  " << max_time_diff_between_msgs);
     rclcpp::QoS qos(10);
     auto rmw_qos_profile = qos.get_rmw_qos_profile();
     static constexpr size_t kImageSynchronizerQueueSize = 10u;
@@ -38,6 +47,7 @@ class StereoSync : public rclcpp::Node {
           right_compressed_img_sub_);
       compressed_image_synchronizer_->registerCallback(
           std::bind(&StereoSync::compressedImageCallback, this, std::placeholders::_1, std::placeholders::_2));
+      compressed_image_synchronizer_->setAgePenalty(max_time_diff_between_msgs);
     } else {
       left_img_sub_.subscribe(this, left_img_topic_, rmw_qos_profile);
       right_img_sub_.subscribe(this, right_img_topic_, rmw_qos_profile);
@@ -45,6 +55,7 @@ class StereoSync : public rclcpp::Node {
           image_sync_policy(kImageSynchronizerQueueSize), left_img_sub_, right_img_sub_);
       image_synchonizer_->registerCallback(
           std::bind(&StereoSync::imageCallback, this, std::placeholders::_1, std::placeholders::_2));
+      image_synchonizer_->setAgePenalty(max_time_diff_between_msgs);
     }
 
     left_img_pub_ = this->create_publisher<sensor_msgs::msg::Image>("/cam0/image_raw", 100);
@@ -52,9 +63,11 @@ class StereoSync : public rclcpp::Node {
   }
 
  private:
-  std::string left_img_topic_, right_img_topic_;
-  std::string left_compressed_img_topic_, right_compressed_img_topic_;
-  bool compressed_image_;
+  std::string left_img_topic_{}, right_img_topic_{};
+  std::string left_compressed_img_topic_{}, right_compressed_img_topic_{};
+  std::string config_file_{};
+  bool compressed_image_{};
+
   rclcpp::Time previous_stamp_;
   message_filters::Subscriber<sensor_msgs::msg::Image> left_img_sub_, right_img_sub_;
   message_filters::Subscriber<sensor_msgs::msg::CompressedImage> left_compressed_img_sub_, right_compressed_img_sub_;
@@ -64,8 +77,8 @@ class StereoSync : public rclcpp::Node {
 
   rclcpp::Publisher<sensor_msgs::msg::Image>::SharedPtr left_img_pub_, right_img_pub_;
 
-  void compressedImageCallback(const sensor_msgs::msg::CompressedImage::ConstSharedPtr left_msg,
-                               const sensor_msgs::msg::CompressedImage::ConstSharedPtr right_msg) {
+  void compressedImageCallback(const sensor_msgs::msg::CompressedImage::ConstSharedPtr& left_msg,
+                               const sensor_msgs::msg::CompressedImage::ConstSharedPtr& right_msg) {
     RCLCPP_INFO_ONCE(this->get_logger(), "Inside stereo compressed image callback");
 
     rclcpp::Time stamp = rclcpp::Time(left_msg->header.stamp.sec, left_msg->header.stamp.nanosec) +
@@ -92,8 +105,8 @@ class StereoSync : public rclcpp::Node {
     previous_stamp_ = stamp;
   }
 
-  void imageCallback(const sensor_msgs::msg::Image::ConstSharedPtr left_msg,
-                     const sensor_msgs::msg::Image::ConstSharedPtr right_msg) {
+  void imageCallback(const sensor_msgs::msg::Image::ConstSharedPtr& left_msg,
+                     const sensor_msgs::msg::Image::ConstSharedPtr& right_msg) {
     RCLCPP_INFO_ONCE(this->get_logger(), "Inside stereo image callback");
 
     rclcpp::Time stamp = rclcpp::Time(left_msg->header.stamp.sec, left_msg->header.stamp.nanosec) +
